+++ arca/backend/__init__.py
@@ -1,9 +1,8 @@
 from .base import BaseBackend
 from .venv import VenvBackend
 from .docker import DockerBackend
-from .current_environment import CurrentEnvironmentBackend, RequirementsStrategy
+from .current_environment import CurrentEnvironmentBackend
 from .vagrant import VagrantBackend
 
 
-__all__ = ["BaseBackend", "VenvBackend", "DockerBackend", "CurrentEnvironmentBackend", "RequirementsStrategy",
-           "VagrantBackend"]
+__all__ = ["BaseBackend", "VenvBackend", "DockerBackend", "CurrentEnvironmentBackend", "VagrantBackend"]