+++ arca/backend/venv.py
@@ -1,13 +1,15 @@
+import shlex
 import shutil
 import subprocess
 from pathlib import Path
+from typing import Optional
 from venv import EnvBuilder
 
 from git import Repo
 
 from arca.exceptions import BuildError, BuildTimeoutError
 from arca.utils import logger
-from .base import BaseRunInSubprocessBackend
+from .base import BaseRunInSubprocessBackend, RequirementsOptions
 
 
 class VenvBackend(BaseRunInSubprocessBackend):
@@ -20,57 +22,71 @@ class VenvBackend(BaseRunInSubprocessBackend):
     There are no extra settings for this backend.
     """
 
-    def get_virtualenv_name(self, requirements_file: Path) -> str:
+    def get_virtualenv_path(self, requirements_option: RequirementsOptions, requirements_hash: Optional[str]) -> Path:
         """
-        Returns a name of the virtualenv that should be used for this repository.
-
-        Either:
-
-        * hash of the requirements file and Arca version
-        * ``no_requirements_file`` if the requirements file doesn't exist.
-
-        :param requirements_file: :class:`Path <pathlib.Path>` to where the requirements file
-            should be in the cloned repository
-
+        Returns the path to the virtualenv the current state of the repository.
         """
-        if requirements_file is None:
-            return "no_requirements_file"
+        if requirements_option == RequirementsOptions.no_requirements:
+            venv_name = "no_requirements"
         else:
-            return self.get_requirements_hash(requirements_file)
+            venv_name = requirements_hash
+
+        return Path(self._arca.base_dir) / "venvs" / venv_name
 
     def get_or_create_venv(self, path: Path) -> Path:
         """
-        Gets the name of the virtualenv from :meth:`get_virtualenv_name`, checks if it exists already,
+        Gets the location of  the virtualenv from :meth:`get_virtualenv_path`, checks if it exists already,
         creates it and installs requirements otherwise. The virtualenvs are stored in a folder based
         on the :class:`Arca` ``base_dir`` setting.
 
         :param path: :class:`Path <pathlib.Path>` to the cloned repository.
         """
-        requirements_file = self.get_requirements_file(path)
-        venv_name = self.get_virtualenv_name(requirements_file)
+        requirements_option, requirements_hash = self.get_requirements_information(path)
 
-        venv_path = Path(self._arca.base_dir) / "venvs" / venv_name
+        venv_path = self.get_virtualenv_path(requirements_option, requirements_hash)
 
         if not venv_path.exists():
             logger.info(f"Creating a venv in {venv_path}")
             builder = EnvBuilder(with_pip=True)
             builder.create(venv_path)
 
-            if requirements_file is not None:
+            shell = False
+            cmd = None
+            cwd = None
+
+            if requirements_option == RequirementsOptions.pipfile:
+                cmd = ["source", (str(venv_path / "bin" / "activate")), "&&",
+                       "pipenv", "install", "--deploy", "--ignore-pipfile"]
+
+                cmd = " ".join(cmd)
+
+                cwd = path / self.pipfile_location
+                shell = True
+            elif requirements_option == RequirementsOptions.requirements_txt:
+                requirements_file = path / self.requirements_location
 
                 logger.debug("Requirements file:")
                 logger.debug(requirements_file.read_text())
                 logger.info("Installing requirements from %s", requirements_file)
 
-                process = subprocess.Popen([str(venv_path / "bin" / "python3"), "-m", "pip", "install", "-r",
-                                            str(requirements_file)],
-                                           stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+                cmd = [str(venv_path / "bin" / "python3"), "-m", "pip", "install", "-r",
+                       shlex.quote(str(requirements_file))]
+
+            if cmd is not None:
+                logger.info("Running Popen cmd %s, with shell %s", cmd, shell)
+
+                process = subprocess.Popen(cmd,
+                                           stdout=subprocess.PIPE,
+                                           stderr=subprocess.PIPE,
+                                           shell=shell,
+                                           cwd=cwd)
 
                 try:
                     out_stream, err_stream = process.communicate(timeout=self.requirements_timeout)
                 except subprocess.TimeoutExpired:
                     process.kill()
-                    shutil.rmtree(venv_path, ignore_errors=True)
+                    logger.warning("The install command timed out, deleting the virtualenv")
+                    shutil.rmtree(str(venv_path), ignore_errors=True)
 
                     raise BuildTimeoutError(f"Installing of requirements timeouted after "
                                             f"{self.requirements_timeout} seconds.")
@@ -83,7 +99,8 @@ class VenvBackend(BaseRunInSubprocessBackend):
                 logger.debug(err_stream)
 
                 if process.returncode:
-                    venv_path.rmdir()
+                    logger.warning("The install command failed, deleting the virtualenv")
+                    shutil.rmtree(str(venv_path), ignore_errors=True)
                     raise BuildError("Unable to install requirements.txt", extra_info={
                         "out_stream": out_stream,
                         "err_stream": err_stream,
@@ -93,7 +110,7 @@ class VenvBackend(BaseRunInSubprocessBackend):
             else:
                 logger.info("Requirements file not present in repo, empty venv it is.")
         else:
-            logger.info(f"Venv already eixsts in {venv_path}")
+            logger.info(f"Venv already exists in {venv_path}")
 
         return venv_path